requirements.txt
streamlit==1.35.0
PyMuPDF==1.23.7
Pillow==10.3.0
pdfplumber==0.10.3
pytesseract==0.3.10
google-generativeai==0.5.4
python-dotenv==1.0.1